Screen Dump: Improve aspect ratio
The Screen Dump DA (based on dead code found in the original
MouseDesk) does a dump of the DHGR screen to an ImageWriter II printer
via an SSC in Slot 1.. But the aspect ratio is was weird.

DHGR is very, very rougly 2:1 v:h. Per the ImageWriter II Technical
Reference manual, the vertical aspect ratio is 72 DPI based on the
spacing of the strikers. The legacy code used a horizontal dot pitch
of 107 DPI ("Semicondensed"); giving a vertical:horizontal aspect of
107:72 or 1.5:1 so it was vertically squashed. Changing to a
horizontal dot pitch of 144 ("Pica") works, but ends up with very
dense printing due to ink bleed. (At least, according to Virtual ][.

Instead, use a horizontal dot pitch of 72 DPI ("Extended") and double
the rows. This ends up with pleasing output.

The output looks better in both Virtual ][ and Apple2TS.