DeskTop: Ignore non-alpha character at start when naming files/volumes
Previously, when renaming an icon or specifying a new volume name with
Format/Erase, the logic would prevent typing a non-alpha character
(e.g. digit, period) if the field was empty. But if it was non-empty,
but the cursor was at the start of the string, non-alpha characters
were still accepted.

Tweak the logic to look at the caret position, not the string length.