DeskTop: Improve use of OA/SA shortcuts for File > Open/Close
DeskTop allows use of OA+SA+O to do "Open and Close" and OA+SA+W to
"Close All". It also allowed holding OA or SA when clicking on the
menu to make selecting "Open" turn into "Open and Close", but this
wasn't applied to "Close Window" and "Close All". Things got extra
weird if the menu was dropped down using the mouse, then OA+O or SA+O
was pressed using the keyboard - this would be treated the same as
OA+SA+O.

Improve this as best we can:

* Make a modified menu click (either OA or SA) prime the menu so that
  selecting Open or Close turns into the "fancy" version.

* Make pressing OA+SA+O or OA+SA+W while the menu is showing be
  treated as the "fancy" version, but not just OA+O, SA+O, OA+W or
  SA+W, unless the menu was "primed".

Fixes #796