DeskTop: When closing window, select volume icon if folder icon absent
When closing a window, if the corresponding icon is present it will be
(1) used as the target of the animation and (2) selected. But if it is
not present, then (1) its volume icon will be used as the target of
the animation, but (2) nothing will be selected.

Make this consistent, and select whatever the animation target is.