DeskTop: Allow keyboard selection of icons in all windows
Previously, keyboard selection of icons - using arrow keys to select
in icon order, tab to select in lexicographical order, or type-down to
select by name - was restricted to the active window and desktop icons
only.

Now that selection works okay in background windows, it makes more
sense to allow this for all icons. If selection moves into another
window it will be activated.