DESKTOP.SYSTEM: Don't trust a partial copy of DeskTop on RAMCard
The launcher would look for DESKTOP/MODULES/DESKTOP on the RAMCard,
and assume that if it was present then a previous invocation had
copied DeskTop to RAMCard. But that's one of the first files copied,
so it was possible that only a partial copy was made, e.g.
missing APPLE.MENU or EXTRAS, which would cause the system to behave
strangely - see issue #801 for a possible example.

Switch the order of file copies so DESKTOP.SYSTEM itself is copied
last, and use that file as the sentinel.