DeskTop: Make Alias now creates .alias file adjacent to original
Previously, the Special > Make Alias command would create the LNK file
in a LINKS/ subdir of DeskTop's folder. This was easiest to implement
at the time, but unlike similar commands in Mac OS and Windows, and
the folder was not easily discoverable.

Now the LNK file is created adjacent to the original, given a ".alias"
suffix, and rename is immediately triggered. The user is then required
to move it somewhere useful and memorable.

One down side is that aliases for volumes can no longer be created.