DeskTop: Prompt to re-insert system disk for Apple Menu items
This worked in A2D 1.1 but regressed in 1.2 when support for placing
arbitrary files in the Apple Menu was added. Add an entry point that
sets a flag to prompt/retry, and use it from the Apple Menu "DA" item
invocation.

(Also, do a spellcheck pass on the test plan.)