Screen Dump: Fix corruption of "File" on menu bar
It appears that initializing and other calls to the Super Serial Card
firmware while 80-column mode is active causes the text screen to be
cleared. MGTK uses that memory for pattern caches which are then
corrupted, leading to mispaints if unless GrafPorts are swapped.

Work around this by saving/restoring the text page (main and aux)
around the SSC usage, and proactively apply the same logic to the
Print Catalog accessory.

Understanding exactly what causes the SSC to do this would be ideal;
I have not spotted the culprit in the firmware. Also note that the
issue only occurs with real SSC firmware; the problem reproduces in
MAME but not in Virtual ][.

Fixes #810