New DA: Screenshot; rename "Screen Dump" to "Print Screen"
The new DA writes out a graphics file named "Screenshot" to the
application directory i.e. where DeskTop.system lives. If there is an
existing file it will be overwritten. It lives in Extras, but it is
most useful when added to the Apple Menu or as a Shortcut.

Also, to make it easier to distinguish this DA (especially in
localized builds), the "Screen Dump" DA was renamed to "Print Screen"