Invoker: Populate $380 same as Bitsy Bye when launching interpreter
The ProDOS 8 Technical Reference Manual documents that when launching
a system file as an interpreter, $2006 has the filename, and $280
should be populated with the path (absolute, or relative to the
current prefix) to the system file itself. This pathname allows a
system program to determine the directory where other needed files may
reside

ProDOS 2.4's Bitsy Bye deviates from this when launching an
interpreter (BASIC.SYSTEM or BASIS.SYSTEM): $280 holds the containing
path and $380 contains the path to the system file.

To restore better compatibility following eafc037, populate $380 as
well when invoking an interpreter in case it is BASIS.SYSTEM. Here's a
behavior comparison, with differences described below:

                Bitsy                DeskTop
BASIS.SYSTEM invoked:
    Prefix      /VOL/                /VOL/
    $2006       <blank>              <blank>
    $280        /VOL/BASIS.SYSTEM    BASIS.SYSTEM [1]
    $380        /VOL/BASIS.SYSTEM    <blank> [2]
File invoked:
    Prefix      /VOL/                /VOL/
    $2006       FILE                 FILE
    $280        /VOL                 /VOL/BASIS.SYSTEM [3]
    $380        /VOL/BASIS.SYSTEM    /VOL/BASIS.SYSTEM

[1] When BASIS.SYSTEM is invoked directly, Bitsy provides an absolute
path, DeskTop provides a relative path. Either are acceptable per the
ProDOS TRM.

[2] When BASIS.SYSTEM is invoked directly, Bitsy populates $380, while
DeskTop does not. This is not documented, so acceptable.

[3] When invoked as an interpreter, Bitsy sets $280 to the containing
path, and DeskTop sets $280 to the path to the system file. Aligning
on Bitsy's behavior for $280 is probably not critical.