Invoker: Populate $280 same as Bitsy Bye when launching BASIS.SYSTEM
ProDOS 2.4's Bitsy Bye populates $280 with the containing path for the
target file when invoking BASIS.SYSTEM as a fallback interpreter. The
documented ProDOS convention is to set $280 to the path to the system
file itself, but Bitsy Bye does its own thing. For better
compatibility, do the same specifically when launching BASIS.SYSTEM,
but not other interpreters.

Updated behavior comparison:

                Bitsy                DeskTop
/VOL/FILE invoked:
    Prefix      /VOL/                /VOL/
    $2006       FILE                 FILE
    $280        /VOL                 /VOL
    $380        /VOL/BASIS.SYSTEM    /VOL/BASIS.SYSTEM
/VOL/SUBDIR/FILE invoked:
    Prefix      /VOL/SUBDIR/         /VOL/SUBDIR/
    $2006       FILE                 FILE
    $280        /VOL/SUBDIR          /VOL/SUBDIR
    $380        /VOL/BASIS.SYSTEM    /VOL/BASIS.SYSTEM

Also note that Bitsy Bye has the high bits set for '/' characters in
the $280 path (i.e. $AF not $2F) and *all* characters in the $380
path. DeskTop populates both paths with high bits clear.