Turn list box implementation into a proper ToolKit (LBTK)
Like previous work for Button controls (BTK) and Line Edit controls
(LETK), this converts the list box implementation from a "lib" into a
ToolKit with an MLI-style interface (List Box ToolKit or LBTK). The
benefits are:

* Clients no longer need their own copy of the code; accessories can
  share the implementation in DeskTop.
* Streamlined API for callers, via macros.
* File dialog overlay size reduced in DeskTop.
* A client could now have two separate list boxes without duplicating
  code. Nothing currently does this, however.

Some things get more complicated. For example, the File Dialog
implementation in DeskTop lives in Main but the new LBTK lives in Aux;
there is already infrastructure to relay the calls from Main to Aux,
but the list box callbacks require bespoke aux-to-main relays for this
case. Fortunately, the File Dialog code already has special cases for
DeskTop.

No user-facing changes as a result of this.