Coding Style: Prefix nested proc names with '_'
This is already done in several places, but apply it to the big
modules (DeskTop, etc). The motivation is to make it easier to tell
when skimming over code whether a declared proc is part of the larger
internal API of a module or whether the usage is scoped. This can make
it easier to spot redundant code and other readablity improvements.

No binary changes.