DeskTop: Rework and simplify Get Info dialog logic
This dialog was implemented in an "interesting" way where each value
was drawn by preparing a string in the initiating proc, then drawing
it by jumping into a generic UI proc. This resulted in a lot of state
updating for no real benefit.

The theory of distinguishing the model and view is fine... but the
whole GET_FILE_INFO struct can be considered the model. The whole
lifecycle of the dialog is reworked. Doing it in a straightforward way
dramatically simplifies the code and saves a bunch of bytes too.

No functional changes. Saves 169 bytes.