DeskTop: Fix error shown after dragging volume onto another volume
In 95e31ee a fix was made to copy a source folder's creation and
modification dates to the destination using GET_FILE_INFO and
SET_FILE_INFO. This logic was also triggered when copying a volume to
a volume or a volume to a folder. In the latter case this was harmless
but incorrect; the contents are copied not the volume itself, so the
dates shouldn't be updated. In the former case it's also incorrect for
the same reason, but it would trigger an alert because SET_FILE_INFO
doesn't work on volumes.

Fix by making the proc early exit if the source is a volume.