DeskTop: Remove unnecessary window/icon association. Saves 19 bytes.
The logic for opening windows needs to associate the new window with
the "parent" icon (folder or volume). If opening from an icon (the
legacy API entry point) this is implicit. If opening from a path (a
newer API entry point) this is done after the window opens by
searching for a matching icon.

The common code had some redundant logic. Rip it out.

No functional changes.