IconTK Clipping: Fix two icon clipping edge cases
The first bug is that if two overlapping windows had aligned right
edges, an icon "underneath" the intersection of the windows would
mispaint, with some pixels of the desktop showing through.
Investigating this issue led to replacing previous workarounds (e.g.
for #182) with the correct logic.

The second bug is that the clipping logic would terminate when a fully
obscured clipping rectangle was found. In the case where an icon is
obscured in the middle by two overlapping windows, the right hand side
wouldn't draw. Handle this case properly.

Part of a series of changes modifying `CalcWindowIntersections` to
improve maintainability and fix edge cases. Left as unsquashed commits
to enable bisecting in case bugs are found.