MGTK: Add MulDiv, retire lib
Since the routine is used so heavily, add `MulDiv` as a call to MGTK.
While it is pure math and not even used by MGTK itself, so many files
were including lib/muldiv.s that it was getting silly.

This pushed MGTK past the space reserved for it ($4000...$8600) so the
upper limit was bumped up by 2 pages to give more room for future
growth.

All clients are updated; most of them can use a dedicated param block
with constants for either the number, numerator or denominator, which
simplifies the call site. Some that use the routine from main memory
continue to place params at $10... like lib/muldiv.s did. And some
places share param blocks across calls, either because they're using
the blocks at $10 or it was just not worth it to convert.

Overall, this saves space... but it's hard to quantify since it moves
things around a lot. For example, DeskTop still pays for the proc in
aux, but no longer in main. Some param blocks move off the zero page
so it's more bytes to populate each input, but can use constants so
there are fewer inputs to populate.

No functional changes.