Selector: Prevent crash after alert is shown from File > Run a Program...
The file dialog uses memory $1600...$1FFF for state (current path,
filenames, etc). If an alert is shown, e.g. when a non-runnable file
is selected, the same memory is used as the "save area" and is
trashed. This is bad since the file dialog is not torn down; a crash
would happen on the next action.

Fix by stashing the file dialog's state in aux memory, which was
fortunately not used.

(A different approach would be to dismiss the dialog, show the alert,
and then re-launch the dialog, but that involves a lot of repainting
and reloading resources, which would make the normal case of actually
launching a file much slower.)

Fixes #832