DeskTop: Rework complex operations on selection to simplify code
Operations on selection, i.e. copy/move and delete, were most
naturally initiated via drag and drop, with other behavior such as
File > Copy To... and File > Delete having to populate the drag/drop
target rather than just provide a path. Residue from the now removed
Lock/Unlock operations, and adjacent operations such as Copy to
RAMCard (used for Shortcuts > Run a Shortcut...) was present, and it
was all a bit of a mess.

Rework all of this have the operations take a target path and simplify
the code around iterating over selected icons. Saves 55 bytes.

As an unintended side effect, this fixes one bug - holding Apple while
dragging a volume icon onto another icon/window would attempt to move
the files instead of copying, and end in a crash. This is now forced
to be a copy.

This is done with an eye towards #838.