IconTK: Improve DragHighlighted call results
The DragHighlighted results were returned a combination of the struct
(the result byte) and A register. Certain results were complex to
decipher, e.g. dragging a file to the desktop was not distinguishable
from a canceled drag or a file drag within the same window with
modifiers.

Define explicit return values and detangle the code somewhat.

This allows DeskTop's code to unify drag/drop result handling between
file and volume icons, making maintenance easier.

The only functional changes are that drags of volume icons on the
desktop with modifiers held down are no no-ops.

Done with an eye towards #838.