DeskTop: Create containing folder when copying volume to another disk
This matches the behavior on classic Mac OS and on the IIgs, at least
when copying to a disk of a different size. On those systems, if you
drag a disk to another disk and they're the same size it does a full
disk copy, but that's annoying and we're not going to do that.

Resolves #846