MGTK: Fix rendering extending past X=511
The logic for calculating save area byte offsets was incorrect and
didn't handle X values greater than or equal to 512, so showing menus
pushed far to the right would corrupt memory. This isn't currently a
concern but could impact future localizations.

While looking for similar logic to leverage, the existing DivMod7 proc
used for setting up fills and a few other places could be used, but
its implementation was unnecessarily tangled. Untangling that allowed
it to be simplified, and the more potential callers updated to use it.

All said, saves 43 byte and leaves the code much simpler.

No functional changes.