LBTK: Fix possible zero page corruption
"I'm not sure how this ever worked!"

The LBTK dispatch copies the param block to the zero page for fast
access. Due to a stray definition, the copy would overwrite its
own pointer. Somehow this wasn't noticed until I was playing
around with a new localization and it went "boom!"

No known repro with current code.