l10n/i18n: Add Bulgarian localization/support
This is based on the Pravetz 8A/8C machines.

The most complicated part of this is the support for Cyrillic text
using the KOI-7 encoding; unlike the Latin script localizations, the
space limits of this 7-bit encoding mean that lowercase Latin is not
available, so any code that performs case transforms needs to be
disabled.

Thanks to Milen Hristov for translations, testing, and guidance!