DeskTop: Enable OK in File > Copy To... when a volume is selected
DeskTop's File > Copy To... allows the user to open a target folder
and click OK to indicate that was where the file should be copied, or
alternatively to just select a target folder and click OK. But this
latter method did not work for volumes. This inconsistency has been
present since "Drives" was added back in 0f9051b.

How the file picker enables the OK button is controlled by two flags -
"is a selection required?" and "is a volume okay?". The second was
considered subservient to the first. This was okay for DeskTop's
Add/Edit a Shortcut (selection required, volumes okay) and for
Selector's Run a Program (selection required, volumes not okay). For
DeskTop's File > Copy To... neither flag was set, but the setting the
second flag wouldn't have done anything anyway.

So rework the flag testing logic so both flags are considered
independently, and set the second flag when File > Copy To... is run.

Resolves #851.