Date & Time: Ensure windows repaint when ProDOS date is first set
On a system without a real-time clock, ProDOS has an empty date/time.
When the Date & Time DA is run on such a system it will set ProDOS's
date/time, which will be used by DeskTop for drawing list views with
"Today" if appropriate. When the date/time is changed in the dialog a
flag is set to trigger a redraw to avoid partial repaint glitches.

In the special case where the DA is run without changing the date/time
from its initial value (the build date), the ProDOS date/time would be
updated but the flag would not be set, so repaint glitches could
occur.

Update the code to set the flag even if the date/time isn't changed by
the user... but be smart enough to only do that if there's actually a
change, to avoid extra repaints.

Also, some flag usage for "dirty date" vs. "dirty settings" introduced
in f13aaa4 was incorrect; that is partially rewound so that settings
save correctly.

Resolves #852