Selector: Fix rendering of entries after the "Copying to RAMCard..." dialog is shown
Regression introduced in 4dc33868ba85542bcef52db5905584d2d7b5a63b -
the introduction of the ListBox ToolKit increased the resident size of
the Selector app, which pushed the overlay address higher. This caused
the "Copying to RAMCard" overlay to trash the in-memory selector list
cache.

The "Run a Program" overlay already trashed the cache, so support code
would reload the list after it was shown. The same thing just needs to
be done after the other overlay now too.

Fixes #866