DeskTop: Disallow creating a shortcut for an alias copied to RAMCard
As discussed in #880 the behavior of copying an alias to a RAMCard is
not ideal - the directory containing the alias would be copied (by
DESKTOP.SYSTEM, SELECTOR or DESKTOP) not the directory containing the
target.

Block this edge case to avoid surprises, the same as is done for
shortcuts for volumes.