Selector: Fix launching "early" shortcuts copied to RAMCard on use
The change in afa22e5 increased the size of the copying overlay, which
caused the first part of the in-memory copy of the selector list to be
trashed. This would cause copies to RAMCard to fail since the path was
not valid.

Rework the API for the overlay so that the path is passed, rather than
a pointer to a path in memory that might be overwritten.

(This is somewhat related to the fix made in 4984ef8, one commit prior
to introducing the regression!)

Fixes #887