DeskTop: Tidy up grafport setting logic for updates
During update events, if drawing window entries the port needs to be
adjusted to exclude the header from the clip rect. This was done
assuming the SetPort call could fail... but it is defined to
unconditionally succeed. Simplify the call site and update the
comments.

Also, rename the various 'Offset...Port' procs to better reflect what
they're doing - adjusting the GrafPort for drawing window entries.

No functional changes.