Start eliminating name shadowing
Name shadowing can lead to subtle issues where a symbol from a parent
scope is unintentionally used. ca65 doesn't have a warning, but for
common cases where symbols are defined with macros we can hack
something in.

This leaves some TODOs for the very common `.params` block definition
macro as that will require more intrusive rearrangement.

Some of the tweaks, e.g. renaming `ok_button` in a few places in lib/
routines, are gross and a better fix is to rethink the scoping in
the includers.

No binary changes.