DeskTop: Fix spatial keyboard navigation
The change in d3da958 moved IconTK's use of the zero page; this was
mostly fine except for one place where DeskTop passed a param block on
the zero page, with an out param. This meant IconTK would fill in the
out param, then restore its zero page usage and overwrite it. It could
have been worse, though.

Move IconTK back and leave some breadcrumbs to avoid this in the
future.