Rework IF_XYZ macros to eliminate use of scopes
The macros used scopes to manage identifiers for the branching (e.g.
from the IF to the ELSE or END_IF). But scopes run afoul of
https://github.com/cc65/cc65/issues/1548 which required extensive use
of z: to get the right addressing mode.

Rework the macros to avoid the use of scopes, and then remove the z:
qualifiers where we can. There is more dark magic involved, but it's
for the greater good. ("The Greater Good!")

No binary changes.