Macros: Add IF_A_EQ_ONE_OF
This is a common case which is awkward to write and handled
inconsistently. Solve it with macros!

Temporarily there are two forms; IF_A_EQ_ONE_OF and XIF_A_EQ_ONE_OF;
the former has BEQs branch to the BNE, the latter has the BEQs branch
just past it. They behave identically but the latter saves cycles.
Both are included in this commit because both forms are used in the
codebase, and this change is binary neutral. A follow-up will unify
them.

No binary changes.