Work towards aligning recursive copy routines (6 of N)
The main chunk of the DeskTop.system and Selector source code is now
identical. This is done by further rearranging and parameterizing
the logic.

One minor behavior change is that if a parent directory of the needs
creation and this fails a disk space check is no longer done in the
Selector. This step is immediately preceded by a disk space check
anyway, so this is probably not observable.