Work towards aligning recursive copy routines (16 of N)
Don't call GET_FILE_INFO on source during iteration when we have a
perfectly good FileEntry.

Copy the code from DeskTop that populates the needed info from the
FileEntry to save an MLI call and the backing work - lots of seeks
etc. Should speed up startup when copying to RAMCard.