DeskTop: Fix locking/unlocking folders
When locking was moved into Get Info in v1.4 it wasn't tested with
directories. The code assumed that a GET_INFO parameter block was
still intact and ready to be used with SET_INFO to change access bits,
but that wasn't the case with directories, since in v1.3 the total
size was calculated.

Make an extra GET_INFO call after the size calculation is complete (or
aborted) so that locking/unlocking the file doesn't trash the file's
info.