Work towards aligning recursive copy routines (30 of N)
Allow retries in DeskTop in directory read edge case.

When reading file entries, DeskTop handles errors and offers to retry.
But it was ignoring errors when reading over the 5-byte gap between
entries at the end of each block and the start of entries in the next
block. This might be observable if the disk was ejected between
directory blocks - entries would be read from the buffered block, then
the read over the 5-byte gap would fail. Since the error was ignored,
the code would eventually loop around and try to read the next entry;
since the disk is not there the user would be prompted to re-insert.
If the disk was indeed re-inserted, the file mark is probably still
before the 5-byte gap; retrying reading the entry would yield garbage.

The common lib/ code would just fail in this case. Since DeskTop
normally retries, make it retry too.