DeskTop: Fix copying of files when disk swapping is required
DeskTop has code in the file copy routine to support swapping disks.
If the destination disk isn't available at one key point it detects
that and and will do extra work to only keep one ProDOS file record
open, and GET_MARK/CLOSE/OPEN/SET_MARK after each read/write, and
prompt for the other disk.

It's tricky to trigger, since you can't have two volume icons for
disks in the same device, so drag/drop can't be used. The original
MouseDesk/DeskTop UI had a copy dialog that let you specify the
source/destination either using a file picker for source and
destination or typing paths, which made it easier, but it can be done
with the simplified File > Copy To... now..

It's also apparently been broken since the 1980s! It detects if you're
in this "exclusive" mode by opening the source file, then opening the
destination file and if that gives you a Volume Not Found error it
sets a flag and prompts you to insert the destination. It then enters
the read/write loop with the swap logic mentioned above. But since you
just inserted the destination disk, the initial source read ends up
coming off the wrong disk!

Fix the logic to not prompt for the destination disk at that point, so
we enter the loop with the correct file open.