Flow Control Macros: Improve syntax
Instead of requiring `IF_A_EQ #12` with a single macro incorporating
the register operand comparison operator as a single macro, introduce
more intelligent IF macro with limited expression parsing, it can be
written as `IF A EQ #12` or even `IF A = #12`. Flag tests are written
as `IF CC`.

Only a single expression is supported, i.e. no grouping via AND/OR or
simply listing multiple flags. Maybe in the future.

The old style of macros are still supported as aliases for the new
ones. This change doesn't touch the rest of the source, acting as a
test for the newfangled parsing via those aliases.

IF_A_EQ_ONE_OF / IF_A_NE_ALL_OF are untouched.

No binary changes.