Flow Control Macros: IN / NOT_IN operators
These replace `IF_A_EQ_ONE_OF ...` / `IF_A_NE_ALL_OF ...` special
cases with logic for `IF A IN ...` / `IF A NOT_IN ...`, but handled
for all of the statements e.g. ELSE_IF, WHILE, etc.

This also introduces a build target (in extras/) to exercise the
macros, as a quick way to verify that no binary changes were
introduced as macros get refactored. It only covers a handful of
things so far (IN / NOT_IN operators, and index register support) but
with more permutatations than are actually in use.

No binary changes.