Tidy up some comparisons
A handful of places were (seemingly) doing "branch if less-or-equal"
or "branch if greater" which require extra branch ops. In some cases
the order of the comparison can be swapped to eliminate the extra ops,
or the overall algorithm can be tweaked. Other cases were more subtle
so update the comments.

No functional changes.