DeskTop: Fix shortcut file corruption following bc19da5
The change in bc19da5 made `text_input_buf` the place where strings
are composed using the new MessageFormat proc, and this is now used
when doing things like painting window headers for the "10 Items"
(etc) strings.

This buffer is also used when adding/editing shortcuts as the field
where shortcut names are typed in - hence the name. Between editing a
shortcut and the new SELECTOR.LIST file being written out the dialog
is closed... which repaints the window headers. So the new name was
trashed. In the most fun case, the string written out string is
overlong and so when reading the file on boot some memory is trashed
and DeskTop crashes. Fun!

Overloading that buffer's usage is probably a bad idea. But here's a
quick fix to prevent the corruption - the name is copied elsewhere.