Macros: Fix unsigned comparison result for ecmp16
The cmp16 (unsigned comparison) and scmp16 (signed comparison) macros
can't be used to detect equality. That's where ecmp16 comes in. Nearly
all uses of ecmp16 are for just equality, and it worked fine. But it
was documented to also yield ordering via the carry bit... but it was
wrong - it compared low bytes before high bytes! A simple fix.

The one place where the ordering is used is in DeskTop's View > by
Date, when comparing 16-bit years years. It turns out the error would
is not currently visible because all plausible ProDOS-8 file dates end
up with the same high byte: $700 is 1792, and $7FF is 2047. Better fix
it now rather than waiting 22 years...