DeskTop: Improve type-down performance
Since the sorted name cache is still valid between keystrokes, only
re-compute it on the first key. Performance is still pretty terrible
with lots of files, but this helps a bit.

Also, for the "flow" macros: make it an error to compare against a
numeric literal. Because typing "IF A = 0" is tempting but almost
certainly wrong.