DeskTop: Improve type-down and sorted view performance
Both of these operations rely on sorting, and a simple but inefficient
sorting algorithm was used - selection sort, which is O(n^2).

It actually isn't that terrible - the file dialog also relies on this
and it can sort a list of a hundred files in under 2 seconds at 1MHz.
But DeskTop has a lot more constant overhead per comparison since
there are many more indirections to the data. For type-down selection,
with a hundred files the sort would take 12 seconds!

Swapping in a Quicksort drops this down to the 2 second range, which
is still not great but a significant improvement.