DeskTop: Fix header displays of K used/available
The change in bc19da5 delegated the conversion of block counts to
kilobytes to the MessageFormat function. But the header values were
converted from blocks to K already, so the values were off.

Delete some code... but also update all the variable names; the change
is smaller than it looks!