Edit test plan, introduce res/tests.hdv
Update and reword a bunch of test plan entries, and introduce a disk
image with a bunch of the pre-work done for tests. This saves the poor
test engineer from having to do a bunch of setup and simplifies the
wording for several tests by just referring to existing files on the
test image.