Flow Control Macros: Allow NOT to prepend expression
A NOT_BETWEEN #'a', #'z' can be rewritten as NOT A BETWEEN #'a', #'z',
and so on for flag tests (e.g. NOT CC), register comparisons (e.g. NOT
A = foo) and list membership (e.g. NOT A IN #1, #2, #3). This may be
clearer in some cases. Probably not. But it justifies some of the
expression parsing complexity.

Multiple NOTs are actually supported, but that's NOT NOT silly.

No binary changes. (Apart from test new cases.)