Flow Control Macros: dedupe the guts of code generation
There are really only two cases - those like "IF" and "RTS_IF" where
the branch is taken if the expression is false, and those like "WHILE"
and "BREAK_IF" where the branch is taken if the expression is true. So
factor those out into more helper macros and streamline everything.

No binary changes.