MGTK: Fix MouseKeys entry/exit testing
Entering and exiting MouseKeys mode was flaky.

There were three places in MGTK checking KBD and clearing KBDSTRB
during CheckEventsImpl:

1: the "should mousekeys be activated/deactivated?" check (guarded by
modifier key test) 2: mouse cursor check (which consumes if the mode
is active) 3: the main check for generating key events

When trying to enter MouseKeys mode (OA+SA+Space): if KBD went high
between (1) and (3), then the keypres would be consumed as a normal
event, and so entering would fail.

When trying to exit MouseKeys mode (Escape): if KBD went high between
(1) and (2) then the keypress would be consumed and dropped, and so
exiting would fail.

The fix is to ensure that the keypress is only consumed in one place
per mode. The check in (1) is now moved to after (3) and made
dependent on it. The checks in (2) and (3) are now exclusive depending
on the mode.

This is a fairly terrifying change - tinkering in the guts of the main
GUI event loop - but seems correct.