Flow Control Macros: Introduce CALL / TAIL_CALL / RETURN
Just like readability of conditional branching and looping can be
improved with the use of higher level flow control macros, we can do
the same with function calls.

`CALL proc` is just `JSR proc`, but you can append expressions
for loading registers and setting flags. For example:

    CALL Y=#kOptionFoo, AX=params_addr, C=1

This will emit an LDY, LDA, LDX, and SEC before the JSR. This makes a
typical function call take fewer lines and makes the parameter usage
more self-documenting when setting up the parameters is really
straightforward.

The same infrastructure can be used with `TAIL_CALL` for `JMP`. Same
benefits, and makes it clearer that this isn't a "GOTO" - it's a
optimized function call.

The above two replace the bespoke family of `param_call` helpers that
codified patterns in the original binary.

And in that spirit, `RETURN` allows replacing `return8` and `return16`
with a more general helper that supports returning a wider variety of
register/flag values.

Outstanding questions:

* How to signal that other registers were set prior to the RETURN
  statement.

* How to signal that N and/or Z are the return values, not the
  register itself.

No binary changes.